requirementsfordocker.txt
absl-py==2.1.0
astunparse==1.6.3
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
deepface==0.0.92
filelock==3.15.4
fire==0.6.0
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
gast==0.6.0
gdown==5.2.0
google-api-core==2.19.2
google-auth==2.34.0
google-cloud-secret-manager==2.16.3
google-pasta==0.2.0
googleapis-common-protos==1.65.0
greenlet==3.0.3
grpc-google-iam-v1==0.13.1
grpcio==1.64.1
grpcio-status==1.62.3
gunicorn==22.0.0
h5py==3.11.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
keras==3.4.1
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
mtcnn==0.1.1
mysqlclient==2.2.4
namex==0.0.8
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
proto-plus==1.24.0
protobuf==4.23.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments==2.18.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
retina-face==0.0.17
rich==13.7.1
rsa==4.9
setuptools==70.3.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.31
termcolor==2.4.0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
